+++ CHANGES.rst
@@ -4,7 +4,12 @@ History
 1.2 (unreleased)
 ----------------
 
-- Nothing changed yet.
+- node >= 1.1 is required by `node.behaviors.suppress_lifecycle_events` support
+  [mamico]
+
+- Backward compatibility with pas.plugins.ldap <= 1.8.1 where LdapProps does not have
+  timeout properties.
+  [mamico]
 
 
 1.1 (2022-10-06)